lifts room muting to the chat engine
