+++ app/src/main/kotlin/app/dapk/st/graph/AppModule.kt
@@ -18,7 +18,6 @@ import app.dapk.st.core.extensions.ErrorTracker
 import app.dapk.st.core.extensions.unsafeLazy
 import app.dapk.st.directory.DirectoryModule
 import app.dapk.st.domain.StoreModule
-import app.dapk.st.domain.room.MutedRoomsStorePersistence
 import app.dapk.st.engine.MatrixEngine
 import app.dapk.st.firebase.messaging.MessagingModule
 import app.dapk.st.home.BetaVersionUpgradeUseCase
@@ -164,7 +163,6 @@ internal class FeatureModules internal constructor(
             context,
             storeModule.value.messageStore(),
             deviceMeta,
-            MutedRoomsStorePersistence(storeModule.value.cachingPreferences)
         )
     }
     val homeModule by unsafeLazy {
