+++ features/messenger/src/main/kotlin/app/dapk/st/messenger/MessengerModule.kt
@@ -10,7 +10,6 @@ import app.dapk.st.domain.application.message.MessageOptionsStore
 import app.dapk.st.engine.ChatEngine
 import app.dapk.st.matrix.common.RoomId
 import app.dapk.st.messenger.state.MessengerState
-import app.dapk.st.domain.room.MutedRoomsStore
 import app.dapk.st.messenger.state.messengerReducer
 
 class MessengerModule(
@@ -18,7 +17,6 @@ class MessengerModule(
     private val context: Context,
     private val messageOptionsStore: MessageOptionsStore,
     private val deviceMeta: DeviceMeta,
-    private val mutedRoomsStore: MutedRoomsStore,
 ) : ProvidableModule {
 
     internal fun messengerState(launchPayload: MessagerActivityPayload): MessengerState {
@@ -29,7 +27,6 @@ class MessengerModule(
                 CopyToClipboard(context.getSystemService(Context.CLIPBOARD_SERVICE) as ClipboardManager),
                 deviceMeta,
                 messageOptionsStore,
-                mutedRoomsStore,
                 RoomId(launchPayload.roomId),
                 launchPayload.attachments,
                 it
