+++ features/messenger/src/main/kotlin/app/dapk/st/messenger/MessengerScreen.kt
@@ -87,16 +87,19 @@ internal fun MessengerScreen(
 
     Column {
         Toolbar(onNavigate = { navigator.navigate.upToHome() }, roomTitle, actions = {
-            OverflowMenu {
-                when (state.isMuted) {
-                    true -> DropdownMenuItem(text = { Text("Unmute notifications", color = MaterialTheme.colorScheme.onSecondaryContainer) }, onClick = {
-                        viewModel.dispatch(ScreenAction.Notifications.Unmute)
-                    })
+            state.roomState.takeIfContent()?.let {
+                OverflowMenu {
+                    when (it.isMuted) {
+                        true -> DropdownMenuItem(text = { Text("Unmute notifications", color = MaterialTheme.colorScheme.onSecondaryContainer) }, onClick = {
+                            viewModel.dispatch(ScreenAction.Notifications.Unmute)
+                        })
 
-                    false -> DropdownMenuItem(text = { Text("Mute notifications", color = MaterialTheme.colorScheme.onSecondaryContainer) }, onClick = {
-                        viewModel.dispatch(ScreenAction.Notifications.Mute)
-                    })
+                        false -> DropdownMenuItem(text = { Text("Mute notifications", color = MaterialTheme.colorScheme.onSecondaryContainer) }, onClick = {
+                            viewModel.dispatch(ScreenAction.Notifications.Mute)
+                        })
+                    }
                 }
+
             }
         })
