+++ matrix/services/room/src/main/kotlin/app/dapk/st/matrix/room/RoomService.kt
@@ -5,10 +5,7 @@ import app.dapk.st.matrix.common.EventId
 import app.dapk.st.matrix.common.RoomId
 import app.dapk.st.matrix.common.RoomMember
 import app.dapk.st.matrix.common.UserId
-import app.dapk.st.matrix.room.internal.DefaultRoomService
-import app.dapk.st.matrix.room.internal.RoomInviteRemover
-import app.dapk.st.matrix.room.internal.RoomMembers
-import app.dapk.st.matrix.room.internal.RoomMembersCache
+import app.dapk.st.matrix.room.internal.*
 
 private val SERVICE_KEY = RoomService::class
 
@@ -27,6 +24,10 @@ interface RoomService : MatrixService {
     suspend fun joinRoom(roomId: RoomId)
     suspend fun rejectJoinRoom(roomId: RoomId)
 
+    suspend fun muteRoom(roomId: RoomId)
+    suspend fun unmuteRoom(roomId: RoomId)
+    suspend fun isMuted(roomId: RoomId): Boolean
+
     data class JoinedMember(
         val userId: UserId,
         val displayName: String?,
@@ -39,6 +40,7 @@ fun MatrixServiceInstaller.installRoomService(
     memberStore: MemberStore,
     roomMessenger: ServiceDepFactory<RoomMessenger>,
     roomInviteRemover: RoomInviteRemover,
+    singleRoomStore: SingleRoomStore,
 ): InstallExtender<RoomService> {
     return this.install { (httpClient, _, services, logger) ->
         SERVICE_KEY to DefaultRoomService(
@@ -46,7 +48,8 @@ fun MatrixServiceInstaller.installRoomService(
             logger,
             RoomMembers(memberStore, RoomMembersCache()),
             roomMessenger.create(services),
-            roomInviteRemover
+            roomInviteRemover,
+            singleRoomStore,
         )
     }
 }
